+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## v7
 
 - APY numbers have been changed to return proper percentages. (v6 value / 100)
+- Added `ExchangeRate` entities.